requirements.txt
Django==5.2.7
django-crispy-forms==2.4
celery==5.5.3
pillow==12.0.0
crispy-bootstrap5==2025.6
djangorestframework==3.16.1
psycopg2==2.9.11
django-environ==0.12.0
gunicorn==23.0.0
psycopg2-binary==2.9.11
